+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/constraint/Query.java
@@ -28,8 +28,6 @@ public interface Query {
 
     Query forAll();
 
-    Query then();
-
     Query then(Rater rater);
 
     Query then(Rating rating);