+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/constraint/QueryI.java
@@ -149,11 +149,6 @@ public class QueryI implements Query {
     }
 
     @Override
-    public Query then() {
-        throw notImplementedYet();
-    }
-
-    @Override
     public Query then(Rater rater) {
         var resultBase
                 = constraint.childrenView().stream()