+++ projects/net.splitcells.gel.core/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 ## [Unreleased]
 ### Major Changes
+* **2022-04-04** Remove no args method `Query.then`,
+  as it does not seem to make sense.
 * **2022-03-31** **\#8**:
   Rename optimizer components,
   so it will be easy to distinguish the offline repair from the planned online repair: